pyproject.toml
[tool.poetry]
name = "simply-eda-app"
version = "0.1.0"
description = "SIMPLY EDA app"
authors = ["pavankumar.bongu@sanpofi.com"]

packages = [
	{ include = "src" }
]


[tool.poetry.dependencies]
python = ">=3.8,<3.12"

streamlit = "^1.0.0"
pandas = "^1.0.0"
numpy = "^1.20.0"
plotly = "^5.0.0"
scikit-learn = "^1.0.0"
snowflake-connector-python = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
